+++ test/unit/com/unifina/controller/CanvasApiControllerSpec.groovy
@@ -1,6 +1,5 @@
 package com.unifina.controller
 
-import com.unifina.ControllerSpecification
 import com.unifina.domain.*
 import com.unifina.service.*
 import com.unifina.signalpath.ModuleException
