+++ test/unit/com/unifina/controller/DashboardItemApiControllerSpec.groovy
@@ -1,6 +1,5 @@
 package com.unifina.controller
 
-import com.unifina.ControllerSpecification
 import com.unifina.domain.*
 import com.unifina.service.DashboardService
 import com.unifina.service.SaveDashboardItemCommand
